+++ packages/components/package.json
@@ -238,9 +238,9 @@
       "./components/hds/form/fieldset/index.js": "./dist/_app_/components/hds/form/fieldset/index.js",
       "./components/hds/form/file-input/base.js": "./dist/_app_/components/hds/form/file-input/base.js",
       "./components/hds/form/file-input/field.js": "./dist/_app_/components/hds/form/file-input/field.js",
-      "./components/hds/form/header/description/index.js": "./dist/_app_/components/hds/form/header/description/index.js",
+      "./components/hds/form/header/description.js": "./dist/_app_/components/hds/form/header/description.js",
       "./components/hds/form/header/index.js": "./dist/_app_/components/hds/form/header/index.js",
-      "./components/hds/form/header/title/index.js": "./dist/_app_/components/hds/form/header/title/index.js",
+      "./components/hds/form/header/title.js": "./dist/_app_/components/hds/form/header/title.js",
       "./components/hds/form/helper-text/index.js": "./dist/_app_/components/hds/form/helper-text/index.js",
       "./components/hds/form/index.js": "./dist/_app_/components/hds/form/index.js",
       "./components/hds/form/indicator/index.js": "./dist/_app_/components/hds/form/indicator/index.js",
@@ -256,9 +256,9 @@
       "./components/hds/form/radio/field.js": "./dist/_app_/components/hds/form/radio/field.js",
       "./components/hds/form/radio/group.js": "./dist/_app_/components/hds/form/radio/group.js",
       "./components/hds/form/section/field-group/index.js": "./dist/_app_/components/hds/form/section/field-group/index.js",
-      "./components/hds/form/section/header/description/index.js": "./dist/_app_/components/hds/form/section/header/description/index.js",
+      "./components/hds/form/section/header/description.js": "./dist/_app_/components/hds/form/section/header/description.js",
       "./components/hds/form/section/header/index.js": "./dist/_app_/components/hds/form/section/header/index.js",
-      "./components/hds/form/section/header/title/index.js": "./dist/_app_/components/hds/form/section/header/title/index.js",
+      "./components/hds/form/section/header/title.js": "./dist/_app_/components/hds/form/section/header/title.js",
       "./components/hds/form/section/index.js": "./dist/_app_/components/hds/form/section/index.js",
       "./components/hds/form/select/base.js": "./dist/_app_/components/hds/form/select/base.js",
       "./components/hds/form/select/field.js": "./dist/_app_/components/hds/form/select/field.js",
